refactor: introduced a new data distributer module
